Compute FBX keytime interpolation factor in double to avoid int64 overflow

+++ code/AssetLib/FBX/FBXConverter.cpp
@@ -3592,7 +3592,9 @@ void FBXConverter::InterpolateKeys(aiVectorKey *valOut, const KeyTimeList &keys,
             const KeyTimeList::value_type timeA = std::get<0>(kfl)->at(id0);
             const KeyTimeList::value_type timeB = std::get<0>(kfl)->at(id1);
 
-            const ai_real factor = timeB == timeA ? ai_real(0.) : static_cast<ai_real>((time - timeA)) / (timeB - timeA);
+            // compute in double: file-controlled int64 keytimes can be extreme
+            // enough that time-timeA or timeB-timeA overflows int64_t
+            const ai_real factor = timeB == timeA ? ai_real(0.) : static_cast<ai_real>((static_cast<double>(time) - static_cast<double>(timeA)) / (static_cast<double>(timeB) - static_cast<double>(timeA)));
             const ai_real interpValue = static_cast<ai_real>(valueA + (valueB - valueA) * factor);
 
             result[std::get<2>(kfl)] = interpValue;
